make getTime() implementation private
